Refactor of TimePeriodParser and TimePeriodParser classes
- Extracted of helper classes to Utilities folder.
