+++ .NET/Microsoft.Recognizers.Text.DateTime/Chinese/Parsers/TimePeriodFunctions.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.Recognizers.Text.DateTime.Utilities;
+using System;
 using System.Text;
 using DateObject = System.DateTime;
